refactor: split youtube.py (1435 lines) into youtube/ package
Split monolithic services/youtube.py (1435 lines) into a youtube/ package
with 8 sub-modules, each < 300 lines:

- __init__.py: re-exports all public symbols for backward compat
- _wrapper.py: lazy imports + YtDlpWrapper singleton (164 lines)
- _format_helpers.py: format selector / subtitle helpers (70 lines)
- _csv_helpers.py: CSV batch download parsing (192 lines)
- _cookie_helpers.py: cookie/browser auth helpers (203 lines)
- _strategy.py: search strategy persistence (72 lines)
- _models.py: YouTubeVideo, BatchResult, EstimateResult (106 lines)
- service.py: YouTubeService class (283 lines)

All existing imports continue to work via the package __init__.py.

Issue: #11
